Add Winner message Display
This display adds a function to javascript to display who won rock paper
scissors as well as what choice the computer made compared to the user.
This also adds a function to capitalize the first letter of a word. I
added this additional function to capitalize the sentence for 'Paper
beats rock' without having to add further logic regarding the selection
of user and computer.