Add additional display box to separate win counts
The win-loss-count being a singular element was causing the setting of
the counts of wins to losses to be more complex. I solved this by adding
a separate p element to display the count of each.
I also added some new div elements for organizational purposes with the
intent to later make flexboxes for win-loss counts.