Improve visuals of user interface
This commit adds a color scheme to the game and sets all text to
sans-serif.
It also resizes all images to 100 x 100 px squares as the original size
of some was overflowing their button containers. I also added Flex
display to buttons so labels are under the images rather than beside
them.
I also in this commit set the buttons to scale 1.1 when hovered over,
as well as changing the cursor to a pointer. This change makes it more
clear that the buttons are able to be interacted with.